Refactored theme toggling code.
